Moved tests to separate project
