+++ Calculator/Calculator.sln
@@ -1,6 +1,8 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Calculator", "Calculator\Calculator.csproj", "{79FECF5B-8F75-49D4-99E7-F5005C9520BD}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Calculator", "Calculator\Calculator.csproj", "{9D9D3D6C-FA72-4DDE-A503-017043F70713}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CalculatorTests", "CalculatorTests\CalculatorTests.csproj", "{FCC5A64B-C9F5-4D54-9597-0FB145FBBA88}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -8,9 +10,13 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{79FECF5B-8F75-49D4-99E7-F5005C9520BD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{79FECF5B-8F75-49D4-99E7-F5005C9520BD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{79FECF5B-8F75-49D4-99E7-F5005C9520BD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{79FECF5B-8F75-49D4-99E7-F5005C9520BD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9D9D3D6C-FA72-4DDE-A503-017043F70713}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9D9D3D6C-FA72-4DDE-A503-017043F70713}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9D9D3D6C-FA72-4DDE-A503-017043F70713}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9D9D3D6C-FA72-4DDE-A503-017043F70713}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FCC5A64B-C9F5-4D54-9597-0FB145FBBA88}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FCC5A64B-C9F5-4D54-9597-0FB145FBBA88}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FCC5A64B-C9F5-4D54-9597-0FB145FBBA88}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FCC5A64B-C9F5-4D54-9597-0FB145FBBA88}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
